requirements.txt
fpdf2>=2.5
PyPDF2>=3.0